# backend/mini-facbook-swagger-api.json
{
	"openapi": "3.0.0",
	"info": {
		"version": "1.0.0",
		"title": "Authentication system",
		"description": "An suthentication system for users."
	},
...
	"paths": {
...
		"/users/resetPassword/{resetToken}": {
			"patch": {
				"tags": ["Users"],
				"summary": "Reset password endpoint",
				"operationId": "resetPassword",
				"requestBody": {
					"description": "User new password and new password confirm",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"password": {
										"type": "string",
										"example": "newPass1234"
									},
									"passwordConfirm": {
										"type": "string",
										"example": "newPass1234"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "successful operation",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"status": {
											"type": "string",
											"example": "success"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Should support password and password confirm. / Token is invalid or has expired. / Password not the same."
					},
					"500": {
						"description": "Error in login. Please log in again."
					}
				}
			}
		},
...
		"/posts": {
			"get": {
				"tags": ["Posts"],
				"summary": "Posts endpoint",
				"operationId": "posts",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"status": {
											"type": "string",
											"example": "success"
										},
										"statusCode": {
											"type": "number",
											"example": 200
										},
										"data": {
											"type": "array",
											"items": {
												"type": "object",
												"$ref": "#/components/schemas/Post"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Invalid token, Please log in again."
					},
					"403": {
						"description": "You do not have permission to perform this action."
					}
				}
			},
			"post": {
				"tags": ["Posts"],
				"summary": "Posts endpoint",
				"operationId": "createPost",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"description": "Post body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"text": {
										"type": "string",
										"example": "post caption like that"
									},
									"image": {
										"type": "form-data"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "successful operation",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"status": {
											"type": "string",
											"example": "success"
										},
										"statusCode": {
											"type": "number",
											"example": 200
										},
										"data": {
											"type": "array",
											"items": {
												"type": "object",
												"$ref": "#/components/schemas/Post"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Invalid token, Please log in again."
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"Post": {
				"type": "object",
				"properties": {
					"text": {
						"type": "string",
						"example": "this is post text"
					},
					"image": {
						"type": "string"
					}
				}
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"servers": [
		{
			"description": "Try it Now",
			"url": "http://localhost:3000/api/v1/"
		}
	]
}